\name{argosfilter-package}
\alias{argosfilter-package}
\alias{argosfilter}
\docType{package}
\title{
Argos locations filter
}
\description{
Functions to filter animal satellite tracking data obtained from Argos. 
It is especially indicated for telemetry studies of marine
animals, where Argos locations are predominantly of low-quality.
}
\details{
\tabular{ll}{
Package: \tab argosfilter\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2007-08-14\cr
License: \tab GPL (GNU Public Licence) version 2 or later\cr
}

}
\author{
Carla Freitas <carla@npolar.no 
Maintainer: Carla Freitas <carla@npolar.no
}

\references{

}

\keyword{ package }
\seealso{
}
\examples{

}
